requirements.txt
python-dotenv
PyJWT
Werkzeug
Flask
gunicorn
db-sqlite3
Flask-RESTful
Flask-Bcrypt
Flask-SQLAlchemy
flask-marshmallow
SQLAlchemy